requirements.txt
requests>=2.28.0
ics>=0.7.2
python-crontab>=2.6.0 